% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddChildG.R
\name{AddChildG}
\alias{AddChildG}
\title{Create a subset of observations containing only children matched to parents/guardians
This function creates a data frame of child-parent/guardian pairs, based on a population distribution of age differences. The distribution used in this function is the log normal.
Two data frames are required. The Children data frame contains the age data, to which the Parent (Guardian) data will be applied.
The minimum and maximum ages of parents can be specified, to minimise the result that there are parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at becoming a parent. The presence of too young and too old parents is tested at the start of the process and an attempt is made to remove initial pairings outside this range. Thus,pre-cleaning the Parent data frame is not required, although pairs outside this age range may be output, due to the distribution used.
The minimum proportion prevents the outcome where most/all people of a particular age, eg. the entire set of 25-year-olds, are parents. The default value is NULL, which assumes that all people of any age can be parents. The defalt value is 0, enabling a pre-cleaned data frame of parents to be used.
An even number of observations is output, which is one child-parent pair.}
\usage{
AddChildG(
  Children,
  ChildIDVariable,
  ChildAgeVariable,
  Parents,
  ParentIDVariable,
  ParentAgeVariable,
  shapeUsed,
  rateUsed,
  MinParentAge = NULL,
  MaxParentAge = NULL,
  MinPropRemain = 0,
  DyadIDValue = NULL,
  HouseholdNumVariable = NULL,
  UserSeed = NULL,
  pValueToStop = 0.01,
  NumIterations = 1e+06
)
}
\arguments{
\item{Children}{A data frame containing observations limited to the children to be matched An age column is required. All children in this data frame will be matched to a parent/guardian.}

\item{ChildIDVariable}{The column number for the ID variable in the Children data frame.}

\item{ChildAgeVariable}{The column number for the Age variable in the Children data frame.}

\item{Parents}{A data frame containing observations limited to parents. An age column is required. This can contain the entire set of people who can be parents, as the assignment is made on age at becoming a parent, not current age. This file can contain the people who can be guardians, as well as parents. This data frame must contain at least the same number of observations as the Children data frame.}

\item{ParentIDVariable}{The column number for the ID variable in the Parent data frame.}

\item{ParentAgeVariable}{The column number for the Age variable in the Parent data frame.}

\item{shapeUsed}{The shape of the distribution of parent ages at the time the child is born. For women, this will commonly be the age at childbirth.
@param rateUsed The rate for the distribution of parent ages at the time the child is born. For women, this will commonly be the age at childbirth.}

\item{MinParentAge}{The youngest age at which a person becomes a parent. The default value is NULL, which permits the user to set no lower bound on age. This variable can still be set if nrow(Children) == nrow(Parents).}

\item{MaxParentAge}{The oldest age at which a person becomes a parent. The default value is NULL, which permits the user to set no upper bound on age. This variable can still be set if nrow(Children) == nrow(Parents).}

\item{MinPropRemain}{The minimum proportion of people, at each age, who are not parents. The default is zero, which may result in all people at a specific age being allocated as parents. This will leave age gaps for any future work, and may not be desirable. If nrow(Children) == nrow(Parents), assigning any value other than 0 will result in an error.}

\item{DyadIDValue}{The starting number for generating a variable that identifies the observations in a child-parent dyad. Must be numeric. If no value is provided, the Dyad ID starts at 1.}

\item{HouseholdNumVariable}{The column name for the household variable. This must be supplied, and in quotes.}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{pValueToStop}{= The primary stopping rule for the function. If this value is not set, the critical p-value of .01 is used.}

\item{NumIterations}{The maximum number of iterations used to construct the coupled data frame. This has a default value of 1000000, and is the stopping rule if the algorithm does not converge.}
}
\description{
The function performs a reasonableness check for child ID, child age, parent ID variable, and household number.
}
