% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddChildLn.R
\name{AddChildLn}
\alias{AddChildLn}
\title{Create a subset of observations containing only children matched to parents/guardians
This function creates a data frame of child-parent/guardian pairs, based on a population distribution of age differences. The distribution used is the log normal
Three data frames are required. The ChildDataFrame contains the age data, to which the two parent/guardian data will be applied. The two data frames for the parent/guardian observations are assumed to be split between mothers and fathers.
The minimum and maximum ages of mothers are required, as this ensures that there are no mothers who were too young (e.g. 11 years) or too old (e.g. 70 years) at childbirth. Thus, the mother data frame does not require pre-cleaning for age.
As the mother data frame is used first, the proportions to assign to fathers and the minimum proportion of women at any age who must not be mothers are required. This ensures that, for example, sole fathers can exist. The minimum proportion prevents the outcome where most/all women of a set age are mothers, leaving few or no women who are not mothers. If desired, this proportion can be set to 0 and no count restriction is set for sampling women into the mothers subset.
An even number of observations is output, which is one child-mother pair and one child-father pair.}
\usage{
AddChildLn(
  ChildDataframe,
  ChildIDVariable,
  ChildAgeVariable,
  meanlogUsed,
  sdlogUsed,
  MotherDataframe,
  MotherIDVariable,
  MotherAgeVariable,
  MinMotherAge,
  MaxMotherAge,
  PropMothers,
  MinPropRemain = 0,
  HouseholdNumVariable,
  UserSeed = NULL,
  pValueToStop = 0.01,
  NumIterations = 1e+06
)
}
\arguments{
\item{ChildDataframe}{A data frame containing observations limited to the children to be matched An age column is required. All children in this data frame will be matched to a parent/guardian.}

\item{ChildIDVariable}{The column number for the ID variable in the ChildDataFrame.}

\item{ChildAgeVariable}{The column number for the Age variable in the ChildDataFrame.}

\item{meanlogUsed}{The mean of the natural log for the distribution of mother ages at childbirth.
@param sdlogUsed The standard deviation of the natural log for the distribution of mother ages at childbirth.}

\item{MotherDataframe}{A data frame containing observations limited to mothers.An age column is required. This can contain the entire set of women who can be mothers, as the assignment is made on age at birth, not current age.}

\item{MotherIDVariable}{The column number for the ID variable in the Recipient data frame.}

\item{MotherAgeVariable}{The column number for the Age variable in the Mother data frame.}

\item{MinMotherAge}{The youngest age at which a woman gives birth.}

\item{MaxMotherAge}{The oldest age at which a woman gives birth.}

\item{PropMothers}{The proportion of children who are assigned to mothers, the remainder will be assigned to fathers.}

\item{MinPropRemain}{The minimum proportion of people, at each age, who are not parents. The default is zero, which may result in all people at a specific age being allocated as parents. This will leave age gaps for any future work, and may not be desirable.}

\item{HouseholdNumVariable}{The column name for the household variable. This must be supplied in quotes.}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{pValueToStop}{= The primary stopping rule for the function. If this value is not set, the critical p-value of .01 is used.}

\item{NumIterations}{The maximum number of iterations used to construct the coupled data frame. This has a default value of 1000000, and is the stopping rule
if the algorithm does not converge.}
}
\description{
The function performs a reasonableness check for child ID, child age, mother ID variable, and household number.
}
