% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddChildLnLoop.R
\name{AddChildLnLoop}
\alias{AddChildLnLoop}
\title{Create a subset of observations containing only children matched to parents/guardians
This function creates a data frame of child-parent/guardian pairs, based on a population distribution of age differences. The distribution used in this function is the log normal. However, the matching is affected by the age structure of the children and parent data frames. The distribution provides a framework upon which to base the matching. The final distribution of age differences, however, may not follow a lognormal distribution.
Two data frames are required. The Children data frame contains the age data, to which the Parent (Guardian) data will be applied.
The minimum and maximum ages of parents must be specified. This ensures that there are no parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at the time the child was born. The presence of too young and too old parents is tested throughout this function. Thus, pre-cleaning the Parent data frame is not required..
The minimum proportion prevents the outcome where most/all people of a particular age, eg. the entire set of 25-year-olds, are parents. The default value is NULL, which assumes that all people of any age can be parents. The defalt value is 0, enabling a pre-cleaned data frame of parents to be used.
An even number of observations is output, which is one child-parent pair.}
\usage{
AddChildLnLoop(
  Children,
  ChildIDVariable,
  ChildAgeVariable,
  Parents,
  ParentIDVariable,
  ParentAgeVariable,
  meanlogUsed,
  sdlogUsed,
  MinParentAge = NULL,
  MaxParentAge = NULL,
  MinPropRemain = 0,
  HouseholdNumVariable = NULL,
  DyadIDValue = NULL,
  UserSeed = NULL
)
}
\arguments{
\item{Children}{A data frame containing observations limited to the children to be matched An age column is required. All children in this data frame will be matched to a parent/guardian.}

\item{ChildIDVariable}{The column number for the ID variable in the Children data frame.}

\item{ChildAgeVariable}{The column number for the Age variable in the Children data frame.}

\item{Parents}{A data frame containing observations limited to parents. An age column is required. This can contain the entire set of people who can be parents, as the assignment is made on age at becoming a parent, not current age. This file can contain the people who can be guardians, as well as parents. This data frame must contain at least the same number of observations as the Children data frame.}

\item{ParentIDVariable}{The column number for the ID variable in the Parent data frame.}

\item{ParentAgeVariable}{The column number for the Age variable in the Parent data frame.}

\item{meanlogUsed}{The mean of the natural log for the distribution of parent ages at the time the child is born. For women, this will commonly be the age at childbirth.
@param sdlogUsed The standard deviation of the natural log for the distribution of parent ages at the time the child is born. For women, this will commonly be the age at childbirth.}

\item{MinParentAge}{The youngest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{MaxParentAge}{The oldest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{MinPropRemain}{The minimum proportion of people, at each age, who are not parents. The default is zero, which may result in all people at a specific age being allocated as parents. This will leave age gaps for any future work, and may not be desirable. If nrow(Children) == nrow(Parents), assigning any value other than 0 will result in an error.}

\item{HouseholdNumVariable}{The column name for the household variable. This must be supplied, and in quotes.}

\item{DyadIDValue}{The starting number for generating a variable that identifies the observations in a child-parent dyad. Must be numeric. If no value is provided, the Dyad ID starts at 1.}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\description{
The function performs a reasonableness check for child ID, child age, parent ID variable, and household number.
}
