% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AdultsNoID}
\alias{AdultsNoID}
\title{Non-partnered synthetic people}
\format{
A data frame of 2,213 rows and 5 variables:
\describe{
    \item{Sex}{SEither Male or Female}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{The age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
}
}
\usage{
AdultsNoID
}
\description{
A subset of people from the Township data frame, aged 20 years and older with a relationship status of "NonPartnered".
}
\keyword{datasets}
