% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllEmployers}
\alias{AllEmployers}
\title{Employers and employees, by industry}
\format{
A data frame of 183 rows and 7 variables:
\describe{
    \item{ANZSIC06}{The code and associated name for each industry}
    \item{BusinessCount}{The random-rounded count of employers in the industry}
    \item{EmployeeCount}{The random-rounded count of employees in the industry}
    \item{minCo}{The minimum number of employers in the industry}
    \item{maxCo}{The maximum number of employers in the industry}
    \item{minStaff}{The minimum number of people employed in the industry}
    \item{maxStaff}{The maximum number of people employed in the industry}
}
}
\source{
Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence. The data has been modified by adding in four additional variables, representing the estimated minimum and maximum counts of businesses and employees.
}
\usage{
AllEmployers
}
\description{
The number of businesses and employees by industry, Timaru District, for 2018.
}
\keyword{datasets}
