% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BadRels}
\alias{BadRels}
\title{Synthetic people restricted to an age range}
\format{
A data frame of 7,568 rows and 8 variables:
\describe{
    \item{Sex}{Either Male or Female}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{The age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
    \item{AgeBand}{The ten-year age band for the age}
    \item{MinAge}{The minimum age in the age band}
    \item{MaxAge}{The maximum age in the age band}
    }
}
\usage{
BadRels
}
\description{
A subset of people from the Township data frame, aged between 20 and 91 years. Age bands, and the associated minimum and maximum ages, have been added.
}
\keyword{datasets}
