% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childrenyes.R
\name{childrenyes}
\alias{childrenyes}
\title{Create child- parent/guardian pairs as many-to-one  using an existing household identifier
This function creates a data frame of child-parent/guardian pairs, based on a distribution of age differences. Multiple children will be matched to the same parent.
Two data frames are required: one for children and one for potential parents.
The minimum and maximum ages of parents must be specified. This ensures that there are no parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at the time the child was born. The presence of too young and too old parents is tested throughout this function. Thus, pre-cleaning the parents data frame is not required.}
\usage{
childrenyes(
  children,
  chlidcol,
  chlagecol,
  numchild = 2,
  twinrate = 0,
  parents,
  paridcol,
  paragecol,
  minparage = NULL,
  maxparage = NULL,
  hhidcol = NULL,
  UserSeed = NULL
)
}
\arguments{
\item{children}{A data frame containing the children to be paired with a parent/guardian.}

\item{chlidcol}{The column number for the ID variable in the children data frame.}

\item{chlagecol}{The column number for the age variable in the children data frame.}

\item{numchild}{The number of children that are required in each household.}

\item{twinrate}{The proportion of the child population who are twins.}

\item{parents}{A data frame containing the potential parents. This data frame must contain at least the same number of observations as the children data frame.}

\item{paridcol}{The column number for the ID variable in the parents data frame.}

\item{paragecol}{The column number for the age variable in the parent data frame.}

\item{minparage}{The youngest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{maxparage}{The oldest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{hhidcol}{The column number for the household identifier variable in the parent data frame}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
A list of three  data frames. $Matched contains the data frame of child-parent matches. $Adults contains any unmatched observations from the parents data frame. $Children contains any unmatched observations from the children data frame. $Adults and/or $Children may be empty data frames.
}
\description{
Create child- parent/guardian pairs as many-to-one  using an existing household identifier
This function creates a data frame of child-parent/guardian pairs, based on a distribution of age differences. Multiple children will be matched to the same parent.
Two data frames are required: one for children and one for potential parents.
The minimum and maximum ages of parents must be specified. This ensures that there are no parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at the time the child was born. The presence of too young and too old parents is tested throughout this function. Thus, pre-cleaning the parents data frame is not required.
}
