% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childyes.R
\name{childyes}
\alias{childyes}
\title{Create child- parent/guardian pairs using an existing household identifier
This function creates a data frame of child-parent/guardian pairs, based on a distribution of age differences. The function will use either a skew normal or normal distribution, depending on whether a skew ("alphaused") parameter is provided. The default value for the skew is 0, and using the default will cause a normal distribution to be used.
Two data frames are required: one for children and one for potential parents.
The minimum and maximum ages of parents must be specified. This ensures that there are no parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at the time the child was born. The presence of too young and too old parents is tested throughout this function. Thus, pre-cleaning the parents data frame is not required.}
\usage{
childyes(
  children,
  chlidcol,
  chlagecol,
  parents,
  paridcol,
  paragecol,
  directxi,
  directomega,
  alphaused = 0,
  minparage = NULL,
  maxparage = NULL,
  hhidcol = NULL,
  userseed = NULL
)
}
\arguments{
\item{children}{A data frame containing the children to be paired with a parent/guardian.}

\item{chlidcol}{The column number for the ID variable in the children data frame.}

\item{chlagecol}{The column number for the age variable in the children data frame.}

\item{parents}{A data frame containing the potential parents. This data frame must contain at least the same number of observations as the children data frame.}

\item{paridcol}{The column number for the ID variable in the parents data frame.}

\item{paragecol}{The column number for the age variable in the parent data frame.}

\item{directxi}{If a skew-normal distribution is used, this is the location value. If the default alphaused value of 0 is used, this defaults to the mean value for the normal distribution.}

\item{directomega}{If a skew-normal distribution is used, this is the scale value. If the default alphaused value of 0 is used, this defaults to the standard deviation value for the normal distribution.}

\item{alphaused}{The skew. If a normal distribution is to be used, this can be omitted as the default value is 0 (no skew).}

\item{minparage}{The youngest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{maxparage}{The oldest age at which a person becomes a parent. The default value is NULL, which will cause the function to stop.}

\item{hhidcol}{The column number for the household identifier variable in the parent data frame}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
A list of three data frames. $Matched contains the data frame of child-parent matches. $Adults contains any unmatched observations from the parents data frame. $Children contains any unmatched observations from the children data frame. $Adults and/or $Children may be empty data frames.
}
\description{
Create child- parent/guardian pairs using an existing household identifier
This function creates a data frame of child-parent/guardian pairs, based on a distribution of age differences. The function will use either a skew normal or normal distribution, depending on whether a skew ("alphaused") parameter is provided. The default value for the skew is 0, and using the default will cause a normal distribution to be used.
Two data frames are required: one for children and one for potential parents.
The minimum and maximum ages of parents must be specified. This ensures that there are no parents who were too young (e.g. 11 years) or too old (e.g. 70 years) at the time the child was born. The presence of too young and too old parents is tested throughout this function. Thus, pre-cleaning the parents data frame is not required.
}
