% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couples.R
\name{couples}
\alias{couples}
\title{Construct pairs of people into couples with a household identifier.
This function creates a data frame of couples, based on a distribution of age differences. The function will use either a skew normal or normal distribution, depending on whether a skew ("alphaused") parameter is provided. The default value for the skew is 0, and using the default will cause a normal distribution to be used.
Two data frames are required. One person from each data frame will be matched, based on the age difference distribution specified. If the data frames are different sizes, the smalldf data frame must be the smaller of the two.In this situation, a random subsample of the largedf data frame will be used.
Both data frames must be restricted to only those people that will have a couples match performed.}
\usage{
couples(
  smalldf,
  smlidcol = NULL,
  smlagecol = NULL,
  largedf,
  lrgidcol = NULL,
  lrgagecol = NULL,
  directxi = NULL,
  directomega = NULL,
  alphaused = 0,
  hhidstart = NULL,
  hhidvar = NULL,
  userseed = NULL,
  ptostop = NULL,
  numiters = 1e+06
)
}
\arguments{
\item{smalldf}{A data frame containing one set of observations to be paired.}

\item{smlidcol}{The column number for the ID variable in the smalldf data frame.}

\item{smlagecol}{The column number for the age variable in the smalldf data frame.}

\item{largedf}{A data frame containing the second set of observations to be paired.}

\item{lrgidcol}{The column number for the ID variable in the largedf data frame.}

\item{lrgagecol}{The column number for the age variable in the largedf data frame.}

\item{directxi}{If a skew-normal distribution is used, this is the location value. If the default alphaused value of 0 is used, this defaults to the mean value for the normal distribution.}

\item{directomega}{If a skew-normal distribution is used, this is the scale value. If the default alphaused value of 0 is used, this defaults to the standard deviation value for the normal distribution.}

\item{alphaused}{The skew. If a normal distribution is to be used, this can be omitted as the default value is 0 (no skew).}

\item{hhidstart}{The starting number for generating the household identifier value that identifies a couple. Must be numeric.}

\item{hhidvar}{The column name for the household variable. This must be supplied in quotes.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{ptostop}{The critical p-value stopping rule for the function. If this value is not set, the critical p-value of .01 is used.}

\item{numiters}{The maximum number of iterations used to construct the coupled data frame. The default value is 1000000, and is the stopping rule if the algorithm does not converge.}
}
\value{
A list of two data frames $Matched contains the data frame of pairs. $Unmatched contains the unmatched observations from largedf. If there are no unmatched people, $Unmatched will be an empty data frame.
}
\description{
Construct pairs of people into couples with a household identifier.
This function creates a data frame of couples, based on a distribution of age differences. The function will use either a skew normal or normal distribution, depending on whether a skew ("alphaused") parameter is provided. The default value for the skew is 0, and using the default will cause a normal distribution to be used.
Two data frames are required. One person from each data frame will be matched, based on the age difference distribution specified. If the data frames are different sizes, the smalldf data frame must be the smaller of the two.In this situation, a random subsample of the largedf data frame will be used.
Both data frames must be restricted to only those people that will have a couples match performed.
}
\examples{
library(dplyr)

set.seed(1)
PartneredFemales <- Township \%>\%
  filter(Sex == "Female", Relationship == "Partnered")
PartneredMales <- Township \%>\%
  filter(Sex == "Male", Relationship == "Partnered")

# partners females and males, using a normal distribution, with the females
being younger by a mean of -2 and a standard deviation of 3
OppSexCouples <- couples(PartneredFemales, smlidcol=3, smlagecol=4,
                         PartneredMales, lrgidcol=3, lrgagecol=4, directxi = -2,
                         directomega = 3, hhidstart = 100, hhidvar="HouseholdID",
                         userseed = 4, ptostop=.01,  numiters=1000000)
}
