% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OppSexCouples.R
\name{Create.OppSexCouples}
\alias{Create.OppSexCouples}
\title{Create a subset of observations containing only opposite-sex couples}
\usage{
Create.OppSexCouples(dataframe)
}
\arguments{
\item{dataframe}{A data frame containing observations limited to one sex and includes an age column.}

\item{ProbSameSex}{The probability of any observation being assigned to a same-sex couple.}

\item{UpWeight}{If TRUE, a subset of ages will be over-sampled.}

\item{UpWeightLowerAge}{The youngest age for the over-sampling. Required if UpWeight is TRUE.}

\item{UpWeightUpperAge}{The oldest age for the over-sampling. Required if UpWeight is TRUE.}

\item{AgeVariableIndex}{The column number of the data frame that contains the ages. Only used if over-sampling is specified. Required if UpWeight is TRUE.}

\item{CoupleIDValue}{The starting number for generating a variable that identifies the observations in a couple. Must be numeric.}
}
\value{
A data frame of an even number of observations for allocation into same-sex couples.
}
\description{
This is a wrapper for randomly sampling observations into opposite-sex couples.
An even number of observations is output, with an appropriate age-difference distribution between the female and male ages in the couples.
}
\examples{

}
