% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Couples.R
\name{Create.SameSexCouples}
\alias{Create.SameSexCouples}
\title{Create a subset of observations containing only same-sex couples}
\usage{
Create.SameSexCouples(x, ProbSameSex = NULL, UpWeight = FALSE,
  UpWeightProp = NULL, UpWeightLowerAge = NULL,
  UpWeightUpperAge = NULL, AgeVariableIndex = NULL)
}
\arguments{
\item{x}{A data frame containing observations limited to one sex and includes an age column.}

\item{ProbSameSex}{The probability of any observation being assigned to a same-sex couple.}

\item{UpWeight}{If TRUE, a subset of ages will be over-sampled.}

\item{UpWeightLowerAge}{The youngest age for the over-sampling. Required if UpWeight is TRUE.}

\item{UpWeightUpperAge}{The oldest age for the over-sampling. Required if UpWeight is TRUE.}

\item{AgeVariableIndex}{The column number of the data frame that contains the ages. Only used if over-sampling is specified. Required if UpWeight is TRUE.}
}
\value{
A data frame of an even number of observations for allocation into same-sex couples.
}
\description{
This is a wrapper for randomly sampling observations into same-sex couples.
It is mainly used for data frames that contain a subset of observations that require over-sampling.
However, it can also be used to generate a simple random sample.
}
