% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empadd.R
\name{empadd}
\alias{empadd}
\title{Create a data frame of people with employers added}
\usage{
empadd(
  employers,
  empid,
  empcount,
  workers,
  wrkid,
  hourscol,
  hoursmin,
  userseed = NULL
)
}
\arguments{
\item{employers}{A data frame containing employer data.}

\item{empid}{The column number for the employer ID.}

\item{empcount}{The column number that provides the number of employees for each employer.}

\item{workers}{A data frame containing the people that must be matched to employers.}

\item{wrkid}{The column number for the unique value that identifies unique people.}

\item{hourscol}{The column number containing the hours worked by each person Must be an ordered factor or numeric. The levels/values must be ascending for hours worked. This is output as an ordered factor.}

\item{hoursmin}{The relevant factor level/number from hourscol representing the workers. Anything lower than this level/number will be treated as unemployed.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
A data frame of the population, with an employer ID attached to each person. Unemployed people will have an employer ID of 0. All columns in the employers data frame, except for the employee counts, are included in the output data frame.
}
\description{
This function creates a data frame of people and matching employers, if employed. Two data frames are required: one for the population and one for the employers.
A numeric or ordered factor for working hours is required. The minimum value for being in employment must be specified. Anyone coded under this value will be treated as unemployed. Thus, pre-cleaning the population data frame is not required.
The employer data frame can be either a summary in the form of the number of employees by employer. The other option is that each row represents a vacancy for an employee. Thus, an employer with 5 employees may be represented as either: a single row with an employee count of 5, or 5 rows with an employee count of 1 in each row.
}
\examples{
library("dplyr")

EmployedPeople <- empadd(EmployerSet, empid = 3, empcount = 2, Township, wrkid = 3,
                         hourscol = 5, hoursmin = 2, userseed = 4)
}
