% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empcreate.R
\name{empcreate}
\alias{empcreate}
\title{Create a data frame of individual employers, each with aggregate employee counts.}
\usage{
empcreate(employers, emptypecol, empnumcol, staffnumcol, userseed = NULL)
}
\arguments{
\item{employers}{A data frame containing aggregate data on employers.}

\item{emptypecol}{The column number for the types of employers. This can be an industry code.}

\item{empnumcol}{The column number containing the aggregate counts of employers by employer type.}

\item{staffnumcol}{The column number containing the aggregate counts of employees by employer type.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
#'A list of three data frames $Companies contains the data frame of synthetic companies, with the number of employees and a mock company name. $Overcounts contains the companies where the number of employers in an industry exceeds the number of employees. This is an informational data frame provides the original values for employee and employee counts. These industries are included in the $Companies file. The count of employers is reset to the count of employees, resulting in synthetic companies with employees count of 1. $NoEmps contains industries with employer counts but no employee counts, employee counts but no employer counts, and a combination of no employers and no employees. These are excluded from the $Companies data frame. Industries with 0 employee and 0 employer counts is likely due to a standardised list of industries being used for all geographic regions. The existence of employers with no employees is indicative of sole-trader/director-only companies. The presence of employees but no employers suggests a data accuracy problem.
}
\description{
This function constructs individual employers from aggregate counts, such as number of employers per employer type. Employer type is often industry, such as "Sheep, Beef Cattle and Grain Farming". Within each employer type, the number of employers is extracted. The number of employees is then randomly assigned to each of those employers, using the total employee count for that industry. A randomisation method is used to ensure that the company counts can be quite dissimilar across the employers within a type. However, this is constructed by the ratio of employers to employees. If the counts are similar, in this case the number of employees will tend to be 1 for each employer.
}
\examples{
library("dplyr")

TownshipEmployment <- empcreate(AllEmployers, emptypecol = 1, empnumcol = 2,
                                staffnumcol = 3, userseed = 4)
}
