% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmployerSet}
\alias{EmployerSet}
\title{Synthetic employers and their employee counts}
\format{
A data frame of 225 rows and 3 variables:
\describe{
    \item{ANZSIC06}{The code and associated name for the industry associated with the employer}
    \item{NumEmployees}{The count of employees for the employer}
    \item{Company}{The name of the employer}
}
}
\usage{
EmployerSet
}
\description{
Synthetic employers and their associated number of employees, randomly constructed using the "AllEmployers" data frame.
}
\keyword{datasets}
