% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GroupInfo}
\alias{GroupInfo}
\title{The proportion of people in a relationship, by age band within sex}
\format{
A data frame of 14 rows and 7 variables
\describe{
    \item{Sex}{Either Male or Female}
    \item{AgeBand}{The 10-year age band}
    \item{MinAge}{The minimum age of the age band}
    \item{MaxAge}{The maximum age of the age band}
    \item{Relationship}{All people are Partnered}
    \item{RelProps}{The proportion of people who have a relationship status of "Partnered"}
    \item{MidPoints}{The median age in the age band}
}
}
\usage{
GroupInfo
}
\description{
The estimated proportion of people in a relationship, by age band within sex, for people aged between 20 and 90 years.
}
\keyword{datasets}
