% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoursfix.R
\name{hoursfix}
\alias{hoursfix}
\title{Reallocates the working hours between people in school and people not in school.
This function reallocates working hours so that adolescents in school work fewer hours than adolescents still in school.
The re-allocation is performed initially for the adolescents still in school. This ensures that the shorter hours worked have a higher probability of being allocated to adolescents who are very unlikely to have longer hours worked. The approach is designed to prevent longer working hours, for example full-time hours, being allocated to adolescents who are still studying.
The data frame must be restricted to only those whose hours can be reallocated.}
\usage{
hoursfix(
  adolescents,
  adlidcol = NULL,
  statuscol = NULL,
  hourscol = NULL,
  hoursmax = NULL,
  UserSeed = NULL
)
}
\arguments{
\item{adolescents}{A data frame containing all adolescents who have working hours.}

\item{adlidcol}{The column number for the unique value that identifies unique adolescents.}

\item{statuscol}{The column number containing the indicator of whether an adolescent is in school or has left school. Can be either an ordered factor or numeric. If this is a factor, factor level 1 must be in-school. If it is a numeric variable, the lowest number must be the in-school value. This is output as an ordered factor.}

\item{hourscol}{The column number containing the hours worked by each adolescent. Must be an ordered factor or numeric. The levels/values must be ascending for hours worked. This is output as an ordered factor.}

\item{hoursmax}{The maximum hours worked by adolescents in-school. Must be the relevant factor level/number from HoursWorked.}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
A data of observations, with working hours reallocated so that an adolesent's working hours is compatible with schooling.
}
\description{
Reallocates the working hours between people in school and people not in school.
This function reallocates working hours so that adolescents in school work fewer hours than adolescents still in school.
The re-allocation is performed initially for the adolescents still in school. This ensures that the shorter hours worked have a higher probability of being allocated to adolescents who are very unlikely to have longer hours worked. The approach is designed to prevent longer working hours, for example full-time hours, being allocated to adolescents who are still studying.
The data frame must be restricted to only those whose hours can be reallocated.
}
\examples{
AdolescentWork <- hoursfix(WorkingAdolescents, adlidcol = 3, statuscol = 6, hourscol = 5, hoursmax = 3, UserSeed = 4)
}
