% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InitialDataframe}
\alias{InitialDataframe}
\title{People in age groups, in the Timaru District}
\format{
A data frame with 46,293 rows and 6 variables:
\describe{
    \item{Sex}{Either Male or Female}
    \item{Age.group}{Age group in five-year age bands}
    \item{Relationship}{Relationship status of the person}
    \item{LowerAge}{The youngest age in the Age.group}
    \item{UpperAge}{The oldest age in the Age.group}
    \item{ID}{The unique identifier for the person}
}
}
\source{
Timaru District 2018 census data (tablecodes 8277 and 8395), sourced from Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
InitialDataframe
}
\description{
A data frame of 46,293 synthetic people. Age groups are present, but not ages.
}
\keyword{datasets}
