% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IntoSchools}
\alias{IntoSchools}
\title{Four person households, with a school status for each person}
\format{
A data frame of 3,052 rows and 8 variables
\describe{
    \item{Sex}{Sex of the person}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{The age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
    \item{SchoolStatus}{The indicator of whether the person is in school (Y) or not (N)}
    \item{HouseholdID}{The household identifier for the person}
    \item{SexCode}{Sex, (F)emale or (M)ale}
}
}
\usage{
IntoSchools
}
\description{
Four-person households, consisting of one parent and three children, with a combination of people in school and not in school. Ages 15 through 18 contain a mixture of people in school and those who have left school. This has been constructed from the Township data frame.
}
\keyword{datasets}
