% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LeftSchool}
\alias{LeftSchool}
\title{School leavers}
\format{
A data frame with 120 rows and 4 variables
\describe{
    \item{YearLeft}{The year for the school leaver count}
    \item{Sex}{The sex for the school leaver count}
    \item{Age}{The age for the school leaver count}
    \item{Total}{The count of adolescents who left school in that year, of that age and sex}
}
}
\source{
Ministry of Education. The Ministry of Education's data are licensed by the Ministry of Education for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
LeftSchool
}
\description{
School leavers in the Canterbury Region, counts by age and sex, for the period 2009 to 2018.
}
\keyword{datasets}
