% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NetworkMatrix}
\alias{NetworkMatrix}
\title{The number of contacts for 5000 person}
\format{
A list of 1,000 integers
}
\usage{
NetworkMatrix
}
\description{
A matrix of 1,000 integers constricted using a Poisson distribution. Each value is the number of contacts for a person.
}
\keyword{datasets}
