% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherno.R
\name{otherno}
\alias{otherno}
\title{Create a match of people into households
This function creates a data frame of household inhabitants, with the specified number of inhabitants.
One data frame, containing the people to match, is required. The use of an age distribution for the matching ensures that an age structure is present in the households. A less correlated age structure can be produced by entering a larger standard deviation.
The output data frame of matches will only contain households of the required size. If the number of rows in the people data frame is not divisible by household size, the overcount will be output to a separate data frame.}
\usage{
otherno(
  people,
  pplidcol,
  pplagecol,
  numppl = NULL,
  sdused,
  hhidstart = 1,
  hhidvar = NULL,
  userseed = NULL,
  ptostop = 0.01,
  numiters = 1e+06
)
}
\arguments{
\item{people}{A data frame containing the people to be matched into households.}

\item{pplidcol}{The column number for the ID variable.}

\item{pplagecol}{The column number for the Age variable.}

\item{numppl}{The household size to be constructed.}

\item{sdused}{The standard deviation of the normal distribution for the distribution of ages in a household.}

\item{hhidstart}{The starting number for generating the household identifier value for showing unique households. Must be numeric.}

\item{hhidvar}{The column name for the household variable. This must be supplied, and in quotes.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{ptostop}{= The primary stopping rule for the function. If this value is not set, the critical p-value of .01 is used.}

\item{numiters}{The maximum number of iterations used to construct the household data frame. This has a default value of 1000000, and is the stopping rule if the algorithm does not converge.}
}
\value{
A list of two data frames $Matched contains the data frame of households containing matched people. All households will be of the specified size. $Unmatched, if populated, contains the people that were not allocated to households. If the number of rows in the people data frame is divisible by the household size required, $Unmatched will be an empty data frame.
}
\description{
Create a match of people into households
This function creates a data frame of household inhabitants, with the specified number of inhabitants.
One data frame, containing the people to match, is required. The use of an age distribution for the matching ensures that an age structure is present in the households. A less correlated age structure can be produced by entering a larger standard deviation.
The output data frame of matches will only contain households of the required size. If the number of rows in the people data frame is not divisible by household size, the overcount will be output to a separate data frame.
}
\examples{
NewHouseholds <- otherno(AdultsNoID, pplidcol = 3, pplagecol = 4, numppl = 3, sdused = 3,
                       hhidstart = 1, hhidvar= "TheHousehold", userseed=4, ptostop = .01,
                        numiters = 50000)
}
