% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otheryes.R
\name{otheryes}
\alias{otheryes}
\title{Create a match of people into existing households
This function creates a data frame of household inhabitants, with the specified number of inhabitants.
Two data frames are required. The 'existing' data frame contains the people already in households. The 'additions' data frame contains the people. The use of an age distribution for the matching ensures that an age structure is present in the households. A less correlated age structure can be produced by entering a larger standard deviation.
The output data frame of matches will only contain households of the required size.}
\usage{
otheryes(
  existing,
  exsidcol,
  exsagecol,
  hhidcol = NULL,
  additions,
  addidcol,
  addagecol,
  numppl = NULL,
  sdused,
  userseed = NULL,
  ptostop = 0.01,
  numiters = 1e+06
)
}
\arguments{
\item{existing}{A data frame containing the people already in households.}

\item{exsidcol}{The column number for the ID variable, for people in the existing data frame.}

\item{exsagecol}{The column number for the Age variable, for people in the existing data frame.}

\item{additions}{A data frame containing the people to be added to the existing households.}

\item{addidcol}{The column number for the ID variable, for people to be added to the existing households.}

\item{addagecol}{The column number for the Age variable, for people in the existing data frame.}

\item{numppl}{The number of people to be added to the household.}

\item{sdused}{The standard deviation of the normal distribution for the distribution of ages in a household.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{ptostop}{= The primary stopping rule for the function. If this value is not set, the critical p-value of .01 is used.}

\item{numiters}{The maximum number of iterations used to construct the household data frame. This has a default value of 1000000, and is the stopping rule if the algorithm does not converge.}

\item{hhidstart}{The starting number for generating the household identifier value for showing unique households. Must be numeric.}

\item{hhidvar}{The column name for the household variable. This must be supplied, and in quotes.}
}
\value{
A list of three data frames $Matched contains the data frame of households containing matched people. All households will be of the specified size. $Existing, if populated, contains the excess people in the existing data frame, who could not be allocated additional people. $Additions, if populated, contains the excess people in the additions data frame who could not be allocated to an existing household.
}
\description{
Create a match of people into existing households
This function creates a data frame of household inhabitants, with the specified number of inhabitants.
Two data frames are required. The 'existing' data frame contains the people already in households. The 'additions' data frame contains the people. The use of an age distribution for the matching ensures that an age structure is present in the households. A less correlated age structure can be produced by entering a larger standard deviation.
The output data frame of matches will only contain households of the required size.
}
\examples{

AdultsID <- IntoSchools \%>\%
filter(Age > 20)

NoHousehold <- Township \%>\%
  filter(Age > 20, Relationship == "NonPartnered", !(ID \%in\% c(AdultsID$ID))) \%>\%
  slice_sample(n = 1500)

OldHouseholds <- otheryes(AdultsID, exsidcol = 3, exsagecol = 4, hhidcol = 7,
                          NoHousehold, addidcol = 3, addagecol = 4, numppl = 2, sdused = 3,
                          userseed=4, ptostop = .01, numiters = 5000)

}
