% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ppl4networks}
\alias{Ppl4networks}
\title{Synthetic people living in the Timaru District}
\format{
A data frame with 1,000 rows and 5 variables:
\describe{
    \item{Sex}{Either Male or Female}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{The age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
}
}
\source{
Timaru District 2018 census data (tablecodes 8277, 8395, and 8460), sourced from Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
Ppl4networks
}
\description{
1000 synthetic people, to match the number of people in the NetworkMatrix data frame.
}
\keyword{datasets}
