% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegionalStructure}
\alias{RegionalStructure}
\title{Sex/Age pyramid for teenagers in the Canterbury Region}
\format{
A data frame with 14 observations and 4 variables:
\describe{
    \item{Sex}{The sex relating to the count}
    \item{Age.group}{String variable of age plus the text " years"}
    \item{Value}{The count of adolescents}
    \item{Age}{The age relating to that count}
}
}
\source{
Canterbury region 2018 census data (tablecode 8277), sourced from Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
RegionalStructure
}
\description{
The number of people, by age and sex, living in the Canterbury region, restricted to ages 13 to 19 years.
}
\keyword{datasets}
