% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolind.R
\name{schoolind}
\alias{schoolind}
\title{Add a variable indicating whether the person is at school, or has left school.
This function creates a data frame with a variable indicating whether the person still attending school or is not attending school. This is an factor with two levels. Pre-cleaning so that only people inside the age range for attending school is not required.
Three data frames are required. The first is the data frame that contains the people ("adolescents") to whom the indicator will be applied. The other two data frames are counts: school leaver counts ("leavers"), and the sex/age pyramid counts ("structure") that apply to the school leaver counts. As cumulative proportions of school leavers are calculated, the leavers data frames must contain multiple years of data. For example, if the minimum school leaving age is 17 and the maximum age is 18, then there must be two years of data in the leavers and structure data frames. The structure data frame contains the sex/age counts for the year of the adolescent data. For example, if the adolescents data frame is based on 2021, then the structure data frame should be the counts for 2021.
The variables specifying sex can be numeric, character, or factor. The sole requirement is that the same code is used in all three data frames. For example, if "F" and "M" are used in the adolescents data frame to denote sex, then "F" and "M" are the codes required in both the leavers and structure data frames. Any number of values can be used, so long as they are unique.}
\usage{
schoolind(
  adolescents,
  adlsxcol = NULL,
  adlagecol = NULL,
  adlyear = NULL,
  minschage = NULL,
  maxschage = NULL,
  leavers,
  lvrsxcol = NULL,
  lvragecol = NULL,
  lvrctcol = NULL,
  lvryearcol = NULL,
  structure,
  strusxcol = NULL,
  struagecol = NULL,
  structcol = NULL,
  stvarname = "Status",
  UserSeed = NULL
)
}
\arguments{
\item{adolescents}{A data frame containing all adolescents who potentially have left school due to their age.}

\item{adlsxcol}{The column number for the variable that contain the codes specifying females and males.}

\item{adlagecol}{The column number for the variable that contains the ages of the adolescents. This must be integer format.}

\item{adlyear}{The year of the adolescents data.}

\item{minschage}{The minimum age of a person, normally a child, can enter school, excluding higher education.}

\item{maxschage}{The maximum age of a person, normally an adolescent, can leave school, excluding higher education.}

\item{leavers}{A data frame containing the counts of the school leavers for each year.}

\item{lvrsxcol}{The column number for the variable that contain the codes specifying females and males.}

\item{lvragecol}{The column number containing the ages for school leavers.}

\item{lvrctcol}{The column number containing the counts for each sex/age combination in the data. This must be in numeric or integer format.}

\item{lvryearcol}{The column number containing the year data for each count.}

\item{structure}{A data frame containing the sex/age pyramid to be used.}

\item{strusxcol}{The column number for the variable that contain the codes specifying females and males.}

\item{struagecol}{The column number containing the individual ages.}

\item{structcol}{The column number containing the counts for each sex/age combination in the data}

\item{stvarname}{The name of the variable to contain the status of the children/adolescents for schooling. The output is "Yes" for those still in school and "No" for those not in school. If not specified, the column name is "Status".}

\item{UserSeed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}
}
\value{
A data frame of an observations, with an added column that contains the school status of each person.
}
\description{
The proportion of adolescents, by age and sex, who have left school is printed to the Console.
}
\examples{
WithSchoolInd <- schoolind(Township, 1, 4, 2018, 5, 18, LeftSchool, 2, 3, 4, 1,
RegionalStructure, 1, 4, 3, "SchoolStatus", UserSeed = 4)
}
