% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SchoolsToUse}
\alias{SchoolsToUse}
\title{Schools and their roll counts}
\format{
A data frame with 266 rows and 5 variables
\describe{
   \item{School.ID}{The numeric ID for the school}
   \item{School.Name}{The name for the school}
   \item{Gender}{Indicator of whether the school is (C)o-ed, (F)emale-only, or (M)ale-only}
   \item{AgeInRoll}{The age of possible students}
   \item{RollCount}{The number of students. The value is 0 if no students that age attend.}
}
}
\source{
The Ministry of Education. The Ministry of Education's data are licensed by the Ministry of Education for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
SchoolsToUse
}
\description{
Nineteen schools in the Canterbury region, with their 2018 roll counts.
}
\keyword{datasets}
