% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SingleAges}
\alias{SingleAges}
\title{Sex/Age pyramid data for Timaru District}
\format{
A data frame with 190 rows and 4 variables:
\describe{
    \item{Age.group}{Age group, in five-year age bands}
    \item{Sex}{Either Male or Female}
    \item{Value}{The number of people that age and sex}
    \item{Age}{Age at last birthday}
}
}
\source{
Timaru District 2018 census data (tablecode 8277), sourced from Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
SingleAges
}
\description{
The number of people, by age and sex, living in the Timaru District.
}
\keyword{datasets}
