% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socnet.R
\name{socnet}
\alias{socnet}
\title{This function creates social networks between people, based on age differences. A data frame of people with ages is required. These are the people who will have social relationships between each other. A a 1x n matrix of counts must also be supplied, where n is the number of rows in the people data frame. As person-to-person pairs are constructed, the sum of the matrix counts must be even. If it is not, the function will randomly select one count from the matrix and add 1 to it. If this correction happens, an explanation, including the index position of the count, will be printed to the console.}
\usage{
socnet(
  people,
  idcol,
  agecol,
  netmax,
  sdused = 0,
  probsame = 0.5,
  userseed = NULL,
  numiters = 1e+06,
  usematrix = "Y"
)
}
\arguments{
\item{people}{A data frame containing people to be matched to each other using social networks.}

\item{idcol}{The column number for each person's unique ID.}

\item{agecol}{The column number for the age variable.}

\item{netmax}{A data frame containing the 1-dimensional matrix of network sizes. Must contain only integers and be the same length as the people data frame.}

\item{sdused}{The standard deviation for the age differences between two people on an edge.}

\item{probsame}{The probability that a friend of a friend is an edge. For example, if A and B and friends, and B and C are friends, this is the probability that C is also a friend of A.}

\item{userseed}{The user-defined seed for reproducibility. If left blank the normal set.seed() function will be used.}

\item{numiters}{The maximum number of iterations used to construct the coupled data frame. This has a default value of 1000000, and is the stopping rule if the algorithm does not converge.}

\item{usematrix}{If an adjacency matrix is output instead of an igraph object. Default is "Y" so an n x n adjacency matrix is output. This is a dgCMatrix.}
}
\value{
Either an igraph of social networks, or a dgCMatrix of n x n.
}
\description{
A normal distribution is used, using the age differences between the pairs. This is centred on 0, i.e. the people in the pair are the same age. If people B and C are in person A's network, probsame is used to determine the likelihood that people B and C know each other. The larger this proportion, the more likely that people in one person's network know each other, compared to random construction of a network between them.
}
\examples{
library("dplyr")
# with the 50\% sample from Township
# output as igraph
NetworksMadeN <- socnet(Ppl4networks, idcol = 3, agecol = 4, NetworkMatrix, sdused=2,
                        probsame = .5, userseed=4, numiters = 100000, usematrix = "N")

# output as n x n matrix
NetworksMadeY <- socnet(Ppl4networks, idcol = 3, agecol = 4, NetworkMatrix, sdused=2,
                        probsame = .5, userseed=4, numiters = 100000, usematrix = "Y")

# smaller sample for visualisation
#
}
