% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Township}
\alias{Township}
\title{Simulated township}
\format{
A data frame with 10,000 rows and 5 variables
\describe{
    \item{Sex}{Sex of the person}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{The age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
}
}
\source{
Timaru District 2018 census data, using tablecodes 8277, 8395, and 8460, sourced from Statistics New Zealand. Statistics New Zealand data are licensed by Stats NZ for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
Township
}
\description{
10,000 simulated people.
}
\keyword{datasets}
