% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorkingAdolescents}
\alias{WorkingAdolescents}
\title{#' People with a school indicator
#'
#' @description
#' A modified version of "Township" data frame with a school indicator variable added.
#'
#' @format A data frame with 10,000 rows and 6 variables
#' \describe{
#'     \item{Sex}{Sex of the person}
#'     \item{Relationship}{Relationship status of the person}
#'     \item{ID}{The unique identifier for the person}
#'     \item{Age}{The age of the person}
#'     \item{HoursWorked}{The number of hours worked in employment, per week. This is an ordered factor.}
#'     \item{SchoolStatus}{The indicator of whether the person is in school (Y) or not (N)}
#' }
"WithSchoolInd"
Adolescents with a school status and employment hours}
\format{
A data frame of 546 observations and 6 variables
\describe{
    \item{Sex}{Sex of the person}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{Age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
    \item{SchoolStatus}{The indicator of whether the person is in school (Y) or not (N)}
}
}
\source{
Timaru District 2018 census data (tablecodes 8277, 8395, and 8460). School status was added using school leavers data produced by the Ministry of Education. Statistics New Zealand and the Ministry of Education's data are licensed, separately, for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
WorkingAdolescents
}
\description{
A set of synthetic adolescents aged between 15 and 18.
}
\keyword{datasets}
