% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorkingAdolescents}
\alias{WorkingAdolescents}
\title{Adolescents with a school status and employment hours}
\format{
A data frame of 478 observations and 6 variables:
\describe{
    \item{Sex}{Either Male or Female}
    \item{Relationship}{Relationship status of the person}
    \item{ID}{The unique identifier for the person}
    \item{Age}{Age of the person}
    \item{HoursWorked}{The number of hours worked in employment, per week}
    \item{SchoolStatus}{The indicator of whether the person is in school (Y) or not (N)}
}
}
\source{
Timaru District 2018 census data (tablecodes 8277, 8395, and 8460). School status was added using school leavers data produced by the Ministry of Education. Statistics New Zealand and the Ministry of Education's data are licensed, separately, for reuse under the Creative Commons Attribution 4.0 International licence.
}
\usage{
WorkingAdolescents
}
\description{
A set of synthetic adolescents aged between 15 and 18.
}
\keyword{datasets}
